Extract handler registry: consolidate handler installation and lookup
Move handler-related functions into dedicated vm/handler.rs module:
- track_run_handler, find_prompt_boundary_by_marker
- handlers_in_caller_chain, chain_entries_in_caller_chain
- find_prompt_boundary_in_caller_chain
- same_effect_python_type, current_handler_chain
- prepare_with_handler (from dispatch.rs)
- should_invoke_handler (from step.rs)

Closes #297
